refactor: add MemoContent component
