+++ build.gradle
@@ -21,7 +21,7 @@ buildscript {
 plugins {
     id 'com.android.application' version '7.4.2' apply false
     id 'com.android.library' version '7.4.2' apply false
-    id 'org.jetbrains.kotlin.android' version '1.8.0' apply false
+    id 'org.jetbrains.kotlin.android' version '1.8.20-RC2' apply false
     id 'com.google.dagger.hilt.android' version '2.44.2' apply false
     id 'org.jetbrains.kotlin.plugin.serialization' version '1.8.0' apply false
 }
